--- sandbox-data/src/main/java/com/exadel/sandbox/dto/UserDto.java
@@ -1,7 +1,0 @@
-package com.exadel.sandbox.dto;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class UserDto {
-}
