+++ sandbox-service/src/main/java/com.exadel.sandbox/service/impl/UserSecurityServiceImpl.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Set;
 
 @Service
-public class UserServiceImpl implements UserDetailsService,UserService {
+public class UserSecurityServiceImpl implements UserDetailsService,UserService {
         @Autowired
         private UserRepository userRepository;
